pyproject.toml
[project]
name = "tradevision"
version = "1.0.0"
description = "AI-Powered Multi-Market Trading Analysis System"
requires-python = ">=3.12"
dependencies = [
    "yfinance>=0.2.36",
    "finnhub-python>=2.4.19",
    "alpaca-py>=0.21.0",
    "pandas>=2.2.0",
    "pandas-ta>=0.3.14b",
    "feedparser>=6.0.11",
    "httpx>=0.27.0",
    "mcp[cli]>=1.0.0",
    "python-dotenv>=1.0.1",
    "apscheduler>=3.10.4",
    "beautifulsoup4>=4.12.0",
    "googlenewsdecoder>=0.1.7",
]

[project.scripts]
trader-mcp = "trader.mcp_server.server:main"
trader-collect-data = "scripts.collect_data:main"
trader-collect-news = "scripts.collect_news:main"
trader-daily-report = "scripts.daily_report:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
